ci: add CI_ONLY_RUN config option for limiting number of builds
If we are composing a large chain of PRs, it's extremely painful to have
to wait for every entry in the build matrix to run. So we add the
CI_ONLY_RUN environment variable which allows us to negatively define a
skip list of bulid configs that will be skipped by exiting 0 early.

Also, limit race tests to only running on master.